Cleaning Code
